Add a crash() method
It displays a pretty Apollo-13 themed error message if SDL is
initialized and we have a reasonable expectation that we can display
error messages without exploding. Failing that, it calls ugly_crash(),
which either complains to stderr or invokes a platform-specific crash
method, such as one that displays a Mac OS X alert dialog.

Crashing ultimately results in the process exiting with a non-zero
exit code. crash() never returns.