Add support for blitting Image objects to Surface objects
Palette support is not yet enabled as this will cause everything that uses
it to be black. We need to more thoroughly convert image blitting to use Image
objects instead. Part of this will be the elimination of the pixels() method
on the Surface object.

Add missing palette copy support