Create display::PalettizedSurface
A Surface is now a generic 2D array of pixels that knows how to
operate on its contents. The game only really has one type of Surface
right now -- a PalettizedSurface -- which is a kind of Surface that
uses 8-bit pixel values pointing into a Palette, supporting optional
transparency.

display::Image no longer exists. In its stead is display::image::*
functions which can read a buffer of PNG data and return a Surface.
display/image/* is just for image I/O.