Make native palettes autoriative
This changeset retires SDLPalette and LegacyPalette, since both of
those store palettes in arbitrary local formats that have to be copied.
Instead, it provides us with SDLPaletteWrapper, allowing access to an
underlying SDL_Palette *, and it modifies Surfaces as appropriate.

Some of the code requires a pal[768] array, so some allowance must be
made. Thus, display::AutoPal: it get()s a PaletteInterface into a
local pal[768] array when constructed, and it set()s that array on
destruction. It's intended to be used in small scopes as a transition
measure.