Update manual.md
The description of the 'config' file is misleading -- it gives an example of editing a line in the file that says to write "random_nauts=1".  That syntax is incorrect and it should be "random_nauts 1" instead.  I've tried with the equals sign and it does not affect the game -- a space must be used instead.