Correct indexing when accessing intel images
Aligns the internal Intel briefing indexing, which starts at zero, with
the indexing of the images, which begins at one. They were misaligned
because the intel_background.png had been assigned the first position.
When generating the filename internally, we cleverly add 1. This is
preferable to renaming all image files and modifying the but2png
utility.