Fix crash when breaking assigned flight crew
Stops the game from crashing when user attempts to break a flight crew
assigned as primary or backup on a schedule mission. Attempting to do so
prompts an dialog telling the user to cancel the mission first. This
caused the overwritten area of the screen to be saved to buffer, but the
buffer had a mismatched palette, resulting in a crash.