Prevent crash when using large casing in VAB
Fixes a graphics error when assigning hardware in the VAB. When
displaying a vehicle hardware mock-up that required a large casing
image (multiperson capsule or multiple illustrated payload components)
the rocket fuel filler image size was calculated incorrectly. This led
to the bottom-right corner being above the top-left corner of the buffer
image, triggering an assertion.