Show proper help messages in Future Missions
Displays the correct help message when trying to assign a crew to a
manned mission when there are no available crews for the selected
spacecraft. Also, the popup screen to assign a spacecraft when planning
a manned future mission would adjust the help text to address the
vehicle selection screen, but didn't change it back when finished.