Fix array overflow error
Removes an array responsible for an overflow error. The array is
intended to track which mission steps had prestige firsts, but is too
short for its purpose. It overflows, and other variables are
overwritten. The array is never referenced after assignment, so it is
removed rather than lengthened.

This resolves #200