Fix news event duplication when loading autosaves
Fixes an error where news events would be repeated when loading an
autosave. The autosave is created at the end of the player turn,
and records the program state at that point. This includes the effects
of the turn's news broadcast, which are applied at the beginning. When a
save is loaded, however, the start-of-turn news broadcast is replayed.
The events of the broadcast were being incorrectly reapplied. This
included:
* Effects such as a 10%/10MB penalty would be be reapplied, forcing the
  player to pay another 10MB if they already paid, or increasing the
  penalty to -20%/20MB.
* A second record of the news entry would be created, so it would appear
  twice in the Viewing Stand. This also would throw off the dates for
  subsequent news broadcasts in the Viewing Stand.