Skip start-of-turn maintenance if loading autosave
Checks if a save is an autosave before performing pre-turn actions. The
autosave is created at the end of a player's turn, and records the game
state to that point. This is the same phenomenon as in issue #204 fixed
by commit 075799eb06031a2b8804533d448b2f885705c067. This stops the game
from incorrectly
* generating extra Intel briefings, threatening an array overrun
* resetting R&D mods for the turn
* changing budget records/estimates
* changing the budget expenditure records
* performing duplicate presidential reviews