Undid my last commit
I just tested actually flying a mission downgraded to unmanned.  It showed as a "U SUBORBITAL" but the mission flew as a manned suborbital, and in fact the astronaut aboard was injured.  So if we want to be able to downgrade to unmanned, we'll have to make some kind of change to the way the game handles missions.