Fix issue where Future mission was not assigned
Reassigns the mission code to a Future Mission after it is cleared
following a failed Hardware/Crew assignment in the Future Missions
screen. The future mission data is modified during the process of
assigning a capsule/minishuttle or crew, so if a later part of the
process fails or is cancelled that modified data must be reset.
Consequently, if Hardware/Crew assignment is not completely successful,
the future mission data is cleared. However, the mission code was not
being restored without changing missions.