Reflect duration changes in FM mission penalty
Updates the mission penalty display in Future Missions to reflect the
current state of the navigation bar. The "Duration Pie" display used for
navigation doubles as the duration setting for variable-length missions,
such as Manned Orbital Duration. When a duration mission is selected,
the mission's duration text and penalty will update each time the
Duration setting changes to accurately reflect the current mission
setup. Fixes issue #246 and implements issue #256.