Reduced time for Jupiter probe to arrive
The game originally set the Jupiter and Saturn probes to take the same length of time to arrive at their respective planets: 7 turns.  That doesn't make sense since Saturn is twice as far away, and that's longer than the Pioneers or Voyagers took to reach Jupiter.  This change reduces it to 4 turns.