Added a code of conduct
This isn't in response to anything anyone here has done, but GitHub recommends that projects establish a code of conduct.

We're an inclusive group, but nowhere here does it say anything like that.  On the off chance someone should become interested in joining a group that would be welcoming and happens to find our site, this might encourage them to give us a try.