Fix Titan Tech transfer for multiple programs
Restores disabled technology transfer when purchasing the two-stage
rocket when the One-stage/Boosters and Three-stage/Mega are above 75%.
Having one of each was not giving any bonus beyond the best of either.
Also, the bug could cause purchase of the Two-stage rocket to run into
the technology transfer code for the Three-stage rocket.