Changed training screens to show "SEASONS"
Basic and Advanced Training formerly showed "TENURE: x", x being the number of seasons.  I changed it to say "TENURE: 1 SEASON" or "TENURE: x SEASONS".  This was inspired by our recent addition of "DAYS" to Astronaut History where it shows the number of days a 'naut has spent in space.