Tiny changes to Efficiency and CIA
The Intelligence screen said things like "The CIA reports that the Soviet Union is developing the Soyuz and rate the reliability..." - reports vs. rate is a tense shift or something, so I changed it to "and rates".
In Hardware Efficiency, if you hit the flag button to see the Prestige Points screen, then click the flag again to return to Efficiency, at the top in the middle, a colon appears after ATTEMPTS.  This change removes the extraneous comma.