Fixed mistaken de\scription of AI strategy
This is discussed in #323.  We had assumed the AI_STRATEGY variable represented the AI's chosen primary capsule strategy and changed the STATS section of the endgame to reflect that.  However the strategy variable is more subtle; it has more to do with the path than the capsule, per Michael.  So this changes the endgame to simply show the number as before, but without identifying the AI as, say, COMPUTER3.