Changed default Soviet player name
It took me a while to find it, but I tracked down where the default player names are defined.  I changed the default Soviet player name from KOROLOV to KOROLYOV, which is a better transliteration.  His name was Королёв, which is usually rendered in English as Korolev because the ё looks like an e (and the game showed it as Korolov), but it's actually pronounced "yo" or "ya", so Korolyov is a better transliteration.