Improved visual cues in Capsule Screen
This commit fixes Issue #348 and implements #173.  Now, in the Capsule Screen if you try to break an empty crew, it tells you the crew is empty, instead of saying it's assigned to a mission.  Crews who are assigned to a mission will show in blue if it's a mission this turn or green if it's next turn.  Primary crews show in a darker shade than backup crews.  Also now, to the left of the picture of the capsule, it will tell if the crew is assigned to a mission, whether it's this turn or next, and whether it's the primary or backup crew.