Improvements to the Help screens
The F1 screens have been edited for improved grammar and readability, and have a number of clarifications (e.g., what do the wings and medal mean in Astronaut History?).  The F2 screens have been made to align much better than the originals.  In a few cases I've added to them to include new features (such as the Delay button in the VAB).