Reinstate <sys/param.h> and <unistd.h> includes
<sys/param.h> is required for MAXPATHLEN. Without this header, the
build fails on Mac OS X 10.6. This was reported to MacPorts in
https://github.com/macports/macports-ports/pull/7449

I don't know what <unistd.h> is needed for but this header include has
been present in every version of SDLMain.m distributed in SDL so I
assume it is needed on some version of Mac OS X and it does no harm to
include it.

Reverts part of ad072b78dfc931bbda0db2ed5fba76ca8adc75df. See #20.