Improvements to launch pad images
The American damaged launch pad image didn't look obviously like a damaged pad, and the Soviet one looked like an operational pad with a rocket on it(!)  I think that one was a wrong file copied and pasted somewhere along the line.  I've replaced them with a more obviously damaged US pad, and a damaged Soviet pad.