Add play-by-mail mode
Bringing back play-by-mail by adding a new game mode that sets the MAIL
global variable to 0 on U.S. turns and 1 on Soviet turns. The main
strategy is that U.S. missions are launched twice; once after the U.S.
turn to determine the outcome, and a second time after the Soviet turn
(reusing the mission events from the U.S. turn) to determine Prestige.
Before the save file is sent to the Soviet side, Data->Prestige and the
U.S. mission counter are restored to their previous states.

At the beginning of the turn, the U.S. player is informed about the
Prestige results of the previous turn and about any Prestige Firsts
acheived by the Soviet side. The Soviet side sees the U.S. Prestige
Firsts during the Soviet mission resolution, i.e., like when playing
against the AI.

Turns are played the usual way, with a file saving dialog appearing
after ending the turn. As the Data struct contains two new fields for
mission resolution, this breaks save game compatibility.