Restore save game compatibility
Fixes the breaking of save game compatibility introduced with PBEM mode.
Since the new fields of the Data struct added at the end are irrelevant in
regular games, we can simply fill Data with what is included in the save
file. We also need to set MAIL explicitely to -1 when loading a regular
game.