Changed launch pad to say damaged instead of destroyed
After the notification (mission review or newscast) saying your launch facility has been destroyed or damaged, the game doesn't distinguish between the two.  The launch pad always claims the pad is destroyed, whereas other buildings say it's damaged.  Given that we can't presently tell which it is, it's better to call a destroyed pad "damaged" than to call a damaged pad "destroyed".  This commit changes the Launch Facility screen to say damaged instead.