Fix VAB display for damaged rockets
Displays the rocket safety factor in red in the VAB when the rocket is
damaged. Due to a bug in 66684f1c6154e17ae727c1d750de26783f6b424a,
launch vehicle damage not being reported properly; the test expected
damage to be stored as a positive quantity but it is stored as a
negative modifier.