Fix mission death prestige in PBEM games
Fixes a bug in the PBEM prestige update after a mission death. Mission
deaths are being tracked in the global death variable, which was not set
during the prestige evaluation of U.S. missions at the Soviet side.
Mission death information is now being extracted from the spResult of
the History record of the mission.