Improves U.S. mission resolution in PBEM games
Removes Soviet missions from the mission ordering during the U.S. turn
in PBEM games. Previously, they were included in the ordering but not
executed because of their hardware not being assembled yet. As there
are cases where the hardware is not reliably cleared for future
missions, this change prevents such issues breaking the entire game.