Remove use of global buffer in Viewing Stand
Replace the use of the global variable 'buffer' with a separate string
when reading in old event cards in the Viewing Stand. Aside from the
usual problems of using an unregulated global storage space, news card
data is varying length and this posed a buffer overflow opportunity.

Also fixed the HOME key in the viewing stand so the news display would
go all the way to the top, instead of the second line.