Use Minishuttle launch failure data
Fixes a bug where mission launch steps involving a minishuttle
defaulted to the capsule failure cases. Some of the standard
failure scenarios for a liftoff (fname F000) reference a capsule; the
minishuttle has an entire alternate set (F100) of failure scenarios.
The step hardware, which is a rocket for the launch step, was being
checked instead of the capsule hardware slot.