Award no Prestige when branching to failure
Fixes several issues in which the game awarded Prestige although the
relevant step failed. An example is the "ENGINE BURN FAILURE, CRAFT
OUT OF CONTROL." failure (code 26) mentioned in #332, which led to
the Manned Orbital prestige being awarded. This commit invalidates any
steps for Prestige evaluation in case of a branch to failure (i.e., code
4, 18, 23, and 26).