Small improvements to N1, Saturn, Nova images in VAB
I've always thought the N1 looked unnecessarily fat in the VAB (and elsewhere).  This change slenderizes it, and makes it look more like the real thing I think.

This also makes modifications to the fins on the Saturn and Nova rockets.  They used to be wider than they are tall, and that didn't look right to me.