Adds PhotoRecon to mission hardware global MH
Ensures that Photo Recon will be assigned to its slot in a mission's MH
array. This contines a fix begun in commit
117dcf6a702c1095ceb2b66d55959166d2a24bf0.

Commit 117dcf6a stops MissionSetup from checking outside the bounds of
the MissionType.Hard array. Because the array does not include a slot
for Photo Recon, it was no longer being assigned to the MH global
variable, whereas previously the array overrun had the side effect of
doing so. This commit automatically assigns Photo Recon to the
appropriate MH slot for every mission, as Photo Recon is always
available.

EVA suits are also automatically assigned to a mission, because the VAB
does not explicitly set MissionType.Hard[Mission_EVA] when assigning
mission hardware, but does guarantee that hardware cannot be assigned
for an EVA mission if the EVA program isn't started.