Fix penalty displays in Mission Control
Fixes a bug where mission penalties are incorrectly displayed in the
Mission Control building. The penalty popup screen displayed the
penalties for the original mission even if a downgraded version was
selected. The "Requirement Penalties" listing did not account for
the extra length of a duration mission when calculating penalties.
Fixes #457.