Award lunar orbital prestige on alternate branch
Moves the lunar orbital prestige from lunar orbital activities to lunar
orbital ins burn (analogous to the manned earth orbital) and make this
step valid for prestige evaluation when on an alternate branch. This means
that a docking failure while travelling to the moon on an LM test mission
will no longer deny the lunar orbital prestige. Also fixes an error in the
prestige calculation that did not take the actual mission steps into
account. Fixes #322.