Fixes to buttons in Intel screen
When I added Home, End, PgUp, and PgDn functionality to the Intel screen it disabled the Continue button (Enter worked, but clicking on it didn't).  This change fixes that.  Also the Home button was bringing you to the second rather than the first entry, and clicking on the down button didn't work quite right.  These issues should all be fixed now.