Added warning re. Kickers
R&D will now warn the player if: they've researched Gemini/Voskhod or XMS-2/Lapot to Max R&D or above, and haven't prototyped Apollo/Soyuz, and haven't started a kicker program yet.  The idea is that if it looks like the player is going to use a spacecraft that requires a kicker, they should get a heads-up that it will be needed.