Soviet newscast now says "minor slowdown" in training
The counterpart to the US newscast "The astronauts have, for an undisclosed sum of money, signed a deal with a major magazine..." says "Cosmonaut training has suffered a slowdown due to power outages at facilities."  That makes it sound like your fellows will take an extra turn to complete Basic Training, but in fact the newscast doesn't do anything.  This commit changes it to say "a minor slowdown" to hopefully give the impression that it won't delay you, without contradicting the audio part of the newscast.