Fix random mission intel weightings
Corrects an error in the algorithm for choosing a fake mission for an
intel report. The number of entries in the possible mission list and
probability weightings did not match for one era. A Saturn Flyby was
added as a new possible mission report. Also, the weighting algorithm
was tweaked to adjust a discrepency.