Adds interface for customizing astro/cosmonauts
Creates a GUI for modifying the stats of the astronaut & cosmonaut
rosters. The interface is accessed through the Astronaut Editing
button in the New Game / Preferences menu. The custom roster is saved
as USER.DAT in the save directory. The game had the prior ability to
play with a custom roster, but did not provide a means for editing it.
This partially implements issue #627.