VAB won't highlight downgraded DM
I just played a game where the DM was downgraded due to a catastrophic failure in docking.  The Safety was downgraded but the Max R&D figure wasn't, so for the rest of the game its Safety showed in yellow to warn me that it can be further researched--but of course, you can't research DMs, so that was just annoying.  This adds a special condition to not highlight DMs.