R&D warns depending on spacecraft
The warning in R&D that your budget is low is now dependent on what spacecraft you have.  If you have a lunar-capable craft (e.g., Gemini, Apollo, XMS-2, Jupiter) it warns at 60 or below.  Otherwise it warns at 20 or below, since satellite and Mercury/Vostok hardware is so cheap.