Fix assert failure in intel reports
Fixes a bug in the intelligence report generation, where the Soviet
side can have a mission code up to 57 (56 for the U.S.) because of
the Soyuz Lunar Landing. However, the code had the two sides mixed
up, which could lead to a crash if the Soviet player had scheduled
a Soyuz L.L. Fixes #675.