Fix manned lunar orbital being awarded to unmanned LM parking
Fixes a bug that awarded the manned lunar orbital prestige when
the successful lunar orbital ins burn was only done by an uncrewed
LM during a LOR mission but the manned capsule didn't make it. The
fix checks whether the hardware of the current pad has a capsule
component. Fixes #678.