Higher EN shortens hospital stays
When a newscast says someone has been injured and sends them to the hospital/infirmary, by default they go there for 2 turns.  Anyone sent has a 25% chance per point above 1 to go for only 1 turn.  Anyone with an EN of 0 has a 25% chance of going to the hospital for 3 turns instead.