Adds settings file to handle audio preferences
Replaces the save-specific audio options with a configuration file
stored in the savegame directory. Associating audio options with a
specific game instance requires players who don't play with the default
settings to constantly reselect them, particularly with PBEM games.
This deprecates the audio settings in the save game files and they will
no longer be checked. This resolves #402.