Repair spaceport animations file
Fixes off-by-one data errors in the spaceport animations file. Animation
cel images have a directory entry, specifying their size (in byes) and
location (offset) in the file, as well as a header with width & height
prior the pixel data. The width value was frequently off by one. This
corrects the file, and removes the fix which had been added to the code.