Fix for #359
Fix the option to return to classic 2x scale (640x400) in config file.
The problem was that read_config_file() search for the variable name and when found, then search for the number (1 or 0). Since the name of this option was "4xscale" it encountered '4' first and returned an error. 
Quick fix was declare the option as "xscale" and now the config file works properly.