build warnings resolved:
- In gamemain.cpp, in function MainLoop, fix for registration of
historic-budget, where the loop returned a negative value for an array
position.
- In admin.cpp, resolved a bug were the Load function could assign more
characters in the savefile-name than allowed.
- In file.cpp and filesystem: updated physfs function calls
	- updated use of PHYSFS_getLastErrorCode instead of PHYSFS_getLastError
	- updated use of PHYSFS_enumerateFiles instead of
	PHYSFS_enumerateFilesCallback
- In gamemain.cpp,  VerifySF function now correctly reviews all Equipment,
before was consistent only for Probes. The function has been renamed to
VerifySafety for clarity.
- In admin.cpp, the LegacyLoad function now has 2 auxiliary functions that
deal with the Equipment structs in order to avoid undetermined behavior.
ShiftMemory is for moving bytes in memory and CheckMSF is for resetting the
MSF field.
- In news_suq.cpp
	- Added a little function to roll a 6die.
	- function Nmod remade, conflict with data.h resolved. Manned and
	Capsule	programs are now properly distingued, (renamed SafetyMod
	for clarity).