Switch focus back to right when hit Recruit
In 'naut recruitment, if you had dismissed some people on the left, then pressed R, or clicked the button, to recruit someone on the right side, focus still stayed on the left.  That is, up and down arrows etc. still moved you up/down on the left side, not the right.  This fixes that.